refactor existing UserDefaults migrations
